Final class mocking not working
