+++ src/main/java/com/resource/TwitterController.java
@@ -9,6 +9,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import twitter4j.Status;
 import twitter4j.TwitterException;
+
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -16,53 +17,47 @@ import javax.ws.rs.core.Response;
 @Produces(MediaType.APPLICATION_JSON)
 @Path("/api/1.0/twitter")
 public class TwitterController {
+
     TweetPostRequest tweetPostRequest;
     SendTweet sendTweet;
     RetrieveTweets retrieveTweets;
     TwitterImpl twitterimpl;
-    private Logger logger= LoggerFactory.getLogger(TwitterController.class);
-    public TwitterController(TweetPostRequest tweetPostRequest, SendTweet sendTweet,RetrieveTweets retrieveTweets,TwitterImpl twitterimpl) {
+
+    private final Logger logger = LoggerFactory.getLogger(TwitterController.class);
+
+    public TwitterController(TweetPostRequest tweetPostRequest, SendTweet sendTweet, RetrieveTweets retrieveTweets, TwitterImpl twitterimpl) {
         this.tweetPostRequest = tweetPostRequest;
         this.sendTweet = sendTweet;
-        this.retrieveTweets=retrieveTweets;
-        this.twitterimpl=twitterimpl;
+        this.retrieveTweets = retrieveTweets;
+        this.twitterimpl = twitterimpl;
     }
-    public TwitterController()
-    {
-        twitterimpl=new TwitterImpl();
+
+    public TwitterController() {
+
     }
+
     @POST
     @Path("/postTweet")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response sendTweet(TweetPostRequest tweetPostRequest) {
         String tweet = tweetPostRequest.getMessage();
-        if (StringUtil.isEmpty(tweet))
-        {
+        if (StringUtil.isEmpty(tweet)) {
             logger.error("Enter a Valid Tweet");
             return Response.status(400, "Please Enter a valid tweet").build();
-        }
-        else
-        {
-            try
-            {
-                SendTweet sendTweet =twitterimpl.getSendTweetObject();
-                Status  status= sendTweet.sendTweets(tweet);
-                if (status.getText().equals(tweet))
-                {
+        } else {
+            try {
+                Status status = twitterimpl.sendTweets(tweet);
+                if (status.getText().equals(tweet)) {
                     logger.info("Tweet Send Successfully");
                     return Response.status(200, "Tweeted Successfully").build();
                 } else {
                     logger.error("Tweet Was Not Done Invalid Request");
                     return Response.status(500, "Request was not completed").build();
                 }
-            }
-            catch (TwitterException e)
-            {
+            } catch (TwitterException e) {
                 logger.error("Tweet Was Not Done Invalid Request");
                 return Response.status(500, "Request Was Not Completed").build();
-            }
-            catch (NullPointerException e)
-            {
+            } catch (NullPointerException e) {
                 logger.error("Status Was Not found so Not Able to tweet");
                 throw new NullPointerException("Not Able To Tweet");
             }
@@ -72,7 +67,6 @@ public class TwitterController {
     @GET
     @Path("/getTimeline")
     public Response getTweets() {
-        RetrieveTweets retrieveTweets =twitterimpl.getRetrieveTweetsObject();
-        return Response.ok(retrieveTweets.fetchLatestTweet()).build();
+        return Response.ok(twitterimpl.fetchLatestTweet()).build();
     }
 }
